Updated Pro Tools version check
Request from AVID: One requirement for integration of your new ARA products into Pro Tools 2024.6 and later is that the products should not be shown in earlier releases of Pro Tools.

Previously we supplied some example code that uses a query of the host version number to determine whether or not to show the plugin. This approach was flawed and could result in the plugin not appearing properly, so we have provided an updated example demonstrating the correct approach. The new example uses a feature query to determine whether the host supports the ARA feature for AAX. This query will indicate that earlier versions of Pro Tools do not support the ARA feature.